Improve page behaviour when minimizing
Fixed-width padding means that the padding on the page is unable to shrink as the viewport shrinks.

All instances of fixed-width padding are replaced with padding of the format max(X vw, Y px). This allows the padding to shrink as the size of the viewport shrinks, down to a minimum of X px.